\name{Decay rate estimates}
\alias{Decay rate estimates}

\title{Decay rate estimates}

\param{k decay rate}
\param{A constant}
\param{half_time}

\description{The built functions are designed to find safe estimates for k decay rate and a constnt parameters, calculate half time, and predict the non-linear model. Input is time.

parameters:
k: decay rate
A: constant
half_time: the half life time of sample

Inputs:
x= time

Output:
y= concentration
k
A
}

