% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_life_calculator.R, R/k_calculator.R
\name{half_life}
\alias{half_life}
\title{Half-Life Calculator}
\usage{
half_life(k)

half_life(k)
}
\arguments{
\item{k}{rate constant}
}
\value{
The half-life of a decaying substance

The half-life of measured substance
}
\description{
Calculate the half-life of decaying substance using reactivity rate constant(k)

Calculate the half-life of measured substance using reactivity rate constant(k)
}
\examples{
half_life(k);
C1 <- 20
C2 <- 10
time <- 5
k_calculator(C1, C2, time) \%>\%
     half_life(k)

t_half <- half_life(2.33);
C1 <- 20
C2 <- 10
time <- 5
k <- k_calculator(C1, C2, time) \%>\%

}
