% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_calculator.R
\name{k_calculator}
\alias{k_calculator}
\title{Reactivity Rate Constant (k) Calculator}
\usage{
k_calculator(C1, C2, time)
}
\arguments{
\item{C1}{initial concentration of substance}

\item{C2}{final concentration of substance}

\item{time}{time period of experiment}
}
\value{
The rate constant(k) for the experiment
}
\description{
Calculate rate constant k from initial concentration, final concentration and time using a first-order reaction equation
}
\examples{
C1 <- 20
C2 <- 10
time <- 5
k <- k_calculator(C1, C2, time)
}
